pyproject.toml
[project]
name = "4sim-taskc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "python-docx>=1.2.0",
]
